requirements.txt
requests
python-dotenv

# Testing dependencies
pytest
pytest-mock
pytest-cov

# Development dependencies
black
flake8
isort